+++ ec/src/models/mod.rs
@@ -1,4 +1,4 @@
-use crate::models::short_weierstrass_jacobian::GroupAffine;
+use crate::AffineCurve;
 use ark_ff::{fields::BitIteratorBE, Field, PrimeField, SquareRootField, Zero};
 
 pub mod bls12;
@@ -9,16 +9,29 @@ pub mod mnt6;
 pub mod short_weierstrass_jacobian;
 pub mod twisted_edwards_extended;
 
-pub trait ModelParameters: Send + Sync + 'static {
+/// Model parameters for an elliptic curve.
+pub trait ModelParameters: Send + Sync + Sized + 'static {
     type BaseField: Field + SquareRootField;
     type ScalarField: PrimeField + SquareRootField + Into<<Self::ScalarField as PrimeField>::BigInt>;
-}
+    type Affine: AffineCurve<BaseField = Self::BaseField, ScalarField = Self::ScalarField>;
 
-pub trait SWModelParameters: ModelParameters {
-    const COEFF_A: Self::BaseField;
-    const COEFF_B: Self::BaseField;
     const COFACTOR: &'static [u64];
     const COFACTOR_INV: Self::ScalarField;
+
+    /// Checks that the current point is in the prime order subgroup given
+    /// the point on the curve.
+    fn is_in_correct_subgroup_assuming_on_curve(item: &Self::Affine) -> bool {
+        item.mul_bits(BitIteratorBE::new(Self::ScalarField::characteristic()))
+            .is_zero()
+    }
+}
+
+/// Model parameters for a Short Weierstrass curve.
+pub trait SWModelParameters:
+    ModelParameters<Affine = short_weierstrass_jacobian::GroupAffine<Self>>
+{
+    const COEFF_A: Self::BaseField;
+    const COEFF_B: Self::BaseField;
     const AFFINE_GENERATOR_COEFFS: (Self::BaseField, Self::BaseField);
 
     #[inline(always)]
@@ -37,21 +50,14 @@ pub trait SWModelParameters: ModelParameters {
         }
         *elem
     }
-
-    fn is_in_correct_subgroup_assuming_on_curve(item: &GroupAffine<Self>) -> bool
-    where
-        Self: Sized,
-    {
-        item.mul_bits(BitIteratorBE::new(Self::ScalarField::characteristic()))
-            .is_zero()
-    }
 }
 
-pub trait TEModelParameters: ModelParameters {
+/// Model parameters for a Twisted Edwards curve.
+pub trait TEModelParameters:
+    ModelParameters<Affine = twisted_edwards_extended::GroupAffine<Self>>
+{
     const COEFF_A: Self::BaseField;
     const COEFF_D: Self::BaseField;
-    const COFACTOR: &'static [u64];
-    const COFACTOR_INV: Self::ScalarField;
     const AFFINE_GENERATOR_COEFFS: (Self::BaseField, Self::BaseField);
 
     type MontgomeryModelParameters: MontgomeryModelParameters<BaseField = Self::BaseField>;
@@ -64,6 +70,7 @@ pub trait TEModelParameters: ModelParameters {
     }
 }
 
+/// Model parameters for a Montgomery curve.
 pub trait MontgomeryModelParameters: ModelParameters {
     const COEFF_A: Self::BaseField;
     const COEFF_B: Self::BaseField;