+++ test-curves/src/bn384_small_two_adicity/g1.rs
@@ -4,7 +4,7 @@ use ark_ec::{
 };
 use ark_ff::{field_new, Zero};
 
-use crate::bn384_small_two_adicity::{Fq, Fr};
+use crate::bn384_small_two_adicity::{Fq, Fr, FR_ONE};
 
 pub type G1Affine = GroupAffine<Parameters>;
 pub type G1Projective = GroupProjective<Parameters>;
@@ -15,6 +15,13 @@ pub struct Parameters;
 impl ModelParameters for Parameters {
     type BaseField = Fq;
     type ScalarField = Fr;
+    type Affine = GroupAffine<Self>;
+
+    /// COFACTOR = 1
+    const COFACTOR: &'static [u64] = &[1];
+
+    /// COFACTOR_INV = COFACTOR^{-1} mod r = 1
+    const COFACTOR_INV: Fr = FR_ONE;
 }
 
 impl SWModelParameters for Parameters {
@@ -23,12 +30,6 @@ impl SWModelParameters for Parameters {
 
     /// COEFF_B = 17
     const COEFF_B: Fq = field_new!(Fq, "17");
-
-    /// COFACTOR = 1
-    const COFACTOR: &'static [u64] = &[0x1];
-
-    /// COFACTOR_INV = COFACTOR^{-1} mod r = 1
-    const COFACTOR_INV: Fr = field_new!(Fr, "1");
 
     /// AFFINE_GENERATOR_COEFFS = (G1_GENERATOR_X, G1_GENERATOR_Y)
     const AFFINE_GENERATOR_COEFFS: (Self::BaseField, Self::BaseField) =